requirements.txt
eql==0.7
PyYAML==5.1.2